+++ t/user.t
@@ -23,30 +23,30 @@ BEGIN {
 ###########
 # invalid field
 my $user;
-eval {
-    $user = pkg('User')->new(login => 'login',
-                             password => 'pwd',
-                             crunk => 'X');
-};
+eval { $user = pkg('User')->new(login => 'login', password => 'pwd', crunk => 'X'); };
 is($@ =~ /constructor args are invalid: 'crunk'/, 1, 'new() - invalid field');
 
 # successes
 ############
-my $admin = pkg('User')->new(login => 'admin',
-                             password => 'whale');
-$user = pkg('User')->new(login => 'arobin',
-                         password => 'gIMp');
+my $admin = pkg('User')->new(
+    login    => 'admin',
+    password => 'whale'
+);
+$user = pkg('User')->new(
+    login    => 'arobin',
+    password => 'gIMp'
+);
 
 # save() tests
 ###############
 # failure 1
-eval {$admin->save()};
+eval { $admin->save() };
 isa_ok($@, 'Krang::User::Duplicate');
 is($@ =~ /This object duplicates/s, 1, 'save() - duplicate check');
 
 # failure 2
 $user->group_ids_push('x');
-eval {$user->save()};
+eval { $user->save() };
 isa_ok($@, 'Krang::User::InvalidGroup');
 is($@ =~ /Invalid group_id in object/s, 1, 'save() - invalid group_id check');
 
@@ -65,8 +65,7 @@ ok($user->user_uuid());
 ##########
 {
     no strict 'subs';
-    is($user->$_, undef, "getter - $_")
-      for qw/email first_name last_name mobile_phone phone/;
+    is($user->$_, undef, "getter - $_") for qw/email first_name last_name mobile_phone phone/;
 }
 is(scalar(@{$user->group_ids}), 1, 'getter - group_ids()');
 
@@ -81,7 +80,7 @@ is($user->$method, $user->user_uuid, 'uuid_meth() is correct');
 $user->group_ids(1, 2, 3);
 $user->save();
 my @gids = $user->group_ids();
-for my $i(0..$#gids) {
+for my $i (0 .. $#gids) {
     is($gids[$i] == $i + 1, 1, "setter - group_ids $i");
 }
 for (qw/email first_name last_name mobile_phone phone/) {
@@ -99,7 +98,7 @@ isa_ok($admin, 'Krang::User', 'find() - login');
 # make sure email is '' for testing
 my $email = $admin->email;
 $admin->email(undef);
-eval {$admin->save};
+eval { $admin->save };
 croak("Very Bad things: $@") if $@;
 my $count = pkg('User')->find(login => 'system', hidden => 1, count => 1);
 
@@ -111,17 +110,17 @@ is(scalar @users, 2, 'find - count');
 
 # revert email field
 $admin->email($email);
-eval {$admin->save};
+eval { $admin->save };
 croak("Very Bad things: $@") if $@;
 
 @users = pkg('User')->find(order_by => 'login');
-my @u = sort {$a->{login} cmp $b->{login}} @users;
+my @u = sort { $a->{login} cmp $b->{login} } @users;
 is($users[0]->login, $u[0]->login, 'find - order_by');
 
 @users = pkg('User')->find(limit => 1);
 is(scalar @users, 1, 'find - limit');
 
-@users = pkg('User')->find(group_ids => [1,2,3], login => 'arobin');
+@users = pkg('User')->find(group_ids => [1, 2, 3], login => 'arobin');
 isa_ok($_, 'Krang::User') for @users;
 is($users[0]->login, 'arobin', 'find - group_ids');
 is(scalar @{$users[0]->group_ids()}, 3, 'group_ids - count');
@@ -130,40 +129,43 @@ is(scalar @{$users[0]->group_ids()}, 3, 'group_ids - count');
 ##########################
 # make sure the admin's username and password are 'admin' and 'whale'
 # preserve values for restoration
-my ($clogin, $cpass) = map {$admin->$_} qw/login password/;
+my ($clogin, $cpass) = map { $admin->$_ } qw/login password/;
 $admin->login('system');
 $admin->password('whale');
-eval {$admin->save();};
+eval { $admin->save(); };
 croak("Won't complete tests bad things have happened: $@") if $@;
 
-is(pkg('User')->check_auth('',''), 0, 'check_auth() - failure 1');
-is(pkg('User')->check_auth('system',''), 0, 'check_auth() - failure 2');
+is(pkg('User')->check_auth('',       ''), 0, 'check_auth() - failure 1');
+is(pkg('User')->check_auth('system', ''), 0, 'check_auth() - failure 2');
 ok(pkg('User')->check_auth('system', 'whale'), 'check_auth() - success');
 
 # revert values
 $admin->login($clogin);
 $admin->{password} = $cpass;
-eval {$admin->save();};
+eval { $admin->save(); };
 croak("Won't complete tests bad things have happened: $@") if $@;
-
 
 # delete() tests
 #################
 # create objects to force a delete failure
-my $site = pkg('Site')->new(preview_path => 'a',
-                            preview_url => 'preview.com',
-                            publish_path => 'b',
-                            url => 'live.com');
+my $site = pkg('Site')->new(
+    preview_path => 'a',
+    preview_url  => 'preview.com',
+    publish_path => 'b',
+    url          => 'live.com'
+);
 $site->save();
 my ($cat) = pkg('Category')->find(site_id => $site->site_id());
-my $template = pkg('Template')->new(category_id => $cat->category_id(),
-                                    content => 'ima template, baby',
-                                    filename => 'tmpl.tmpl');
+my $template = pkg('Template')->new(
+    category_id => $cat->category_id(),
+    content     => 'ima template, baby',
+    filename    => 'tmpl.tmpl'
+);
 $template->save();
 $template->checkout();
 
 # failure
-eval {$admin->delete()};
+eval { $admin->delete() };
 isa_ok($@, 'Krang::User::Dependency');
 
 # remove cause of failure
@@ -172,15 +174,14 @@ $template->delete();
 # remove leftover site
 $site->delete();
 
-my $old_user = $user; # save this so we can cleanup later
+my $old_user = $user;    # save this so we can cleanup later
+
 END {
+
     # clean up the user and any old passwords
     foreach my $u ($user, $old_user, $admin) {
-        dbh()->do(
-            'DELETE FROM old_password WHERE user_id = ?',
-            {},
-            $u->user_id
-        );
+        dbh()->do('DELETE FROM old_password WHERE user_id = ?', {}, $u->user_id);
+
         # success
         is($u->delete(), 1, 'delete()') unless $u->hidden;
     }
@@ -188,8 +189,10 @@ END {
 
 # if a scheduled send alert refers to a user in it's context stash
 # then it should be deleted when the user is deleted
-$user = pkg('User')->new(login => 'asdfasdf',
-                         password => 'gIMp');
+$user = pkg('User')->new(
+    login    => 'asdfasdf',
+    password => 'gIMp'
+);
 $user->group_ids_push(1);
 $user->save();
 my $date = localtime();
@@ -205,7 +208,7 @@ my $schedule1 = pkg('Schedule')->new(
     action      => 'send',
     repeat      => 'never',
     date        => $date,
-    context     => [ user_id => $user->user_id ],
+    context     => [user_id => $user->user_id],
 );
 isa_ok($schedule1, pkg('Schedule'));
 $schedule1->save();
@@ -221,7 +224,7 @@ $schedule2->save();
 
 $user->delete();
 my ($schedule) = pkg('Schedule')->find(schedule_id => $schedule1->schedule_id);
-ok(! defined $schedule, 'schedule1 was deleted');
+ok(!defined $schedule, 'schedule1 was deleted');
 ($schedule) = pkg('Schedule')->find(schedule_id => $schedule2->schedule_id);
 isa_ok($schedule, pkg('Schedule'), 'schedule2 was not deleted');
 $schedule->delete();
